refactor: rename ServiceViewModel to ServiceListModel
