+++ DesktopApplicationTemplate.UI/ViewModels/MainViewModel.cs
@@ -3,225 +3,27 @@ using System;
 using System.Collections.ObjectModel;
 using System.Collections.Generic;
 using System.ComponentModel;
-using System.Runtime.CompilerServices;
-using System.Linq;
 using System.IO;
-using System.Windows.Controls;
-using System.Windows.Input;
+using System.Linq;
+using System.Threading.Tasks;
 using System.Windows.Data;
-using System.Text.RegularExpressions;
-using WpfBrush = System.Windows.Media.Brush;
+using System.Windows.Input;
 using WpfBrushes = System.Windows.Media.Brushes;
-using DesktopApplicationTemplate.UI.Services;
-using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.Core.Models;
-using System.Text.Json.Serialization;
-using DesktopApplicationTemplate.UI;
+using DesktopApplicationTemplate.Core.Services;
+using DesktopApplicationTemplate.UI.Services;
 using DesktopApplicationTemplate.UI.Helpers;
 
 namespace DesktopApplicationTemplate.UI.ViewModels
 {
-    public class ServiceViewModel : ViewModelBase
-    {
-        public string DisplayName { get; set; } = string.Empty;
-        public string ServiceType { get; set; } = string.Empty;
-        [JsonIgnore] public Page? Page { get; set; }
-        public int Order { get; set; }
-
-        private WpfBrush _backgroundColor = WpfBrushes.LightGray;
-        public WpfBrush BackgroundColor
-        {
-            get => _backgroundColor;
-            set { _backgroundColor = value; OnPropertyChanged(); }
-        }
-
-        private WpfBrush _borderColor = WpfBrushes.Gray;
-        public WpfBrush BorderColor
-        {
-            get => _borderColor;
-            set { _borderColor = value; OnPropertyChanged(); }
-        }
-        [JsonIgnore] public Page? ServicePage { get; set; }
-
-        public ObservableCollection<string> AssociatedServices { get; } = new();
-
-        private double _totalExecutionTimeMs;
-        private int _executionCount;
-
-        /// <summary>
-        /// Gets the average execution time in milliseconds for operations performed by this service.
-        /// </summary>
-        public double? AverageExecutionTimeMs => _executionCount == 0 ? null : _totalExecutionTimeMs / _executionCount;
-
-        /// <summary>
-        /// Gets or sets the accumulated execution time in milliseconds.
-        /// </summary>
-        public double TotalExecutionTimeMs
-        {
-            get => _totalExecutionTimeMs;
-            set
-            {
-                _totalExecutionTimeMs = value;
-                OnPropertyChanged(nameof(AverageExecutionTimeMs));
-            }
-        }
-
-        /// <summary>
-        /// Gets or sets the number of execution samples recorded for this service.
-        /// </summary>
-        public int ExecutionCount
-        {
-            get => _executionCount;
-            set
-            {
-                _executionCount = value;
-                OnPropertyChanged(nameof(AverageExecutionTimeMs));
-            }
-        }
-
-        /// <summary>
-        /// TCP-specific configuration for this service, if applicable.
-        /// </summary>
-        public TcpServiceOptions? TcpOptions { get; set; }
-
-        /// <summary>
-        /// FTP server-specific configuration for this service, if applicable.
-        /// </summary>
-        public FtpServerOptions? FtpOptions { get; set; }
-
-        /// <summary>
-        /// HTTP-specific configuration for this service, if applicable.
-        /// </summary>
-        public HttpServiceOptions? HttpOptions { get; set; }
-
-        /// <summary>
-        /// HID-specific configuration for this service, if applicable.
-        /// </summary>
-        public HidServiceOptions? HidOptions { get; set; }
-
-        /// <summary>
-        /// Heartbeat-specific configuration for this service, if applicable.
-        /// </summary>
-        public HeartbeatServiceOptions? HeartbeatOptions { get; set; }
-
-        /// <summary>
-        /// File Observer-specific configuration for this service, if applicable.
-        /// </summary>
-        public FileObserverServiceOptions? FileObserverOptions { get; set; }
-
-        /// <summary>
-        /// SCP-specific configuration for this service, if applicable.
-        /// </summary>
-        public ScpServiceOptions? ScpOptions { get; set; }
-        /// <summary>
-        /// CSV creator-specific configuration for this service, if applicable.
-        /// </summary>
-        public CsvServiceOptions? CsvOptions { get; set; }
-
-
-
-        public static Func<string, string, ServiceViewModel?>? ResolveService { get; set; }
-
-
-        private bool _isActive;
-        public bool IsActive
-        {
-            get => _isActive;
-            set
-            {
-                if (_isActive != value)
-                {
-                    _isActive = value;
-                    OnPropertyChanged();
-                    if (_isActive)
-                        AddLog("[Service Activated]", WpfBrushes.Green);
-                    else
-                        AddLog("[Service Deactivated]", WpfBrushes.Red);
-                    ActiveChanged?.Invoke(_isActive);
-                }
-            }
-        }
-
-        public ObservableCollection<LogEntry> Logs { get; set; } = new();
-        public event Action<bool>? ActiveChanged;
-
-        public event Action<ServiceViewModel, LogEntry>? LogAdded;
-
-        public void AddLog(string message, WpfBrush? color = null, LogLevel level = LogLevel.Debug, bool checkReference = true)
-        {
-            var ts = DateTime.Now.ToString("MM.dd.yyyy - HH:mm:ss.fffffff");
-            var entry = new LogEntry { Message = $"{ts} {message}", Color = (color ?? WpfBrushes.Black).ToString(), Level = level };
-            Logs.Insert(0, entry);
-            LogAdded?.Invoke(this, entry);
-            if (checkReference)
-            {
-                HandleReference(message, color ?? WpfBrushes.Black, level);
-            }
-        }
-
-        /// <summary>
-        /// Records a single execution duration for this service and updates the running average.
-        /// </summary>
-        /// <param name="duration">The execution duration to record.</param>
-        /// <exception cref="ArgumentException">Thrown when <paramref name="duration"/> is negative.</exception>
-        public void RecordExecutionTime(TimeSpan duration)
-        {
-            if (duration < TimeSpan.Zero)
-                throw new ArgumentException("Duration must be non-negative", nameof(duration));
-
-            _totalExecutionTimeMs += duration.TotalMilliseconds;
-            _executionCount++;
-            OnPropertyChanged(nameof(AverageExecutionTimeMs));
-        }
-
-        private void HandleReference(string message, WpfBrush color, LogLevel level)
-        {
-            var m = Regex.Match(message, @"^([^.]+)\.([^.]+)\.(.+)$");
-            if (m.Success && ResolveService != null)
-            {
-                var type = m.Groups[1].Value;
-                var name = m.Groups[2].Value;
-                var msg = m.Groups[3].Value;
-                var target = ResolveService(type, name);
-                if (target != null && target != this)
-                {
-                    if (!AssociatedServices.Contains(target.DisplayName))
-                        AssociatedServices.Add(target.DisplayName);
-                    if (!target.AssociatedServices.Contains(DisplayName))
-                        target.AssociatedServices.Add(DisplayName);
-                    target.AddLog(msg, color, level, false);
-                }
-            }
-        }
-
-        public void SetColorsByType()
-        {
-            (BackgroundColor, BorderColor) = ServiceType switch
-            {
-                "TCP" => (WpfBrushes.LightBlue, WpfBrushes.DarkBlue),
-                "HTTP" => (WpfBrushes.LightGreen, WpfBrushes.DarkGreen),
-                "File Observer" => (WpfBrushes.LightSalmon, WpfBrushes.DarkSalmon),
-                "HID" => (WpfBrushes.LightYellow, WpfBrushes.Goldenrod),
-                "Heartbeat" => (WpfBrushes.LightPink, WpfBrushes.DeepPink),
-                "SCP" => (WpfBrushes.LightCyan, WpfBrushes.CadetBlue),
-                "MQTT" => (WpfBrushes.LightGoldenrodYellow, WpfBrushes.Goldenrod),
-                "FTP Server" or "FTP" => (WpfBrushes.LightSteelBlue, WpfBrushes.SteelBlue),
-                _ => (WpfBrushes.LightGray, WpfBrushes.Gray)
-            };
-            OnPropertyChanged(nameof(BackgroundColor));
-            OnPropertyChanged(nameof(BorderColor));
-        }
-    }
-
-
     public class MainViewModel : ViewModelBase
     {
-        public ObservableCollection<ServiceViewModel> Services { get; set; } = new();
+        public ObservableCollection<ServiceListModel> Services { get; set; } = new();
         public ICollectionView FilteredServices { get; }
         public FilterViewModel Filters { get; } = new();
         public ObservableCollection<LogEntry> AllLogs { get; } = new();
-        private ServiceViewModel? _selectedService;
-        public ServiceViewModel? SelectedService
+        private ServiceListModel? _selectedService;
+        public ServiceListModel? SelectedService
         {
             get => _selectedService;
             set
@@ -230,13 +32,13 @@ namespace DesktopApplicationTemplate.UI.ViewModels
                 OnPropertyChanged();
                 LogViewModel.SetLogs(_selectedService?.Logs ?? AllLogs);
                 (RemoveServiceCommand as AsyncRelayCommand)?.RaiseCanExecuteChanged();
-                (EditServiceCommand as RelayCommand<ServiceViewModel?>)?.RaiseCanExecuteChanged();
+                (EditServiceCommand as RelayCommand<ServiceListModel?>)?.RaiseCanExecuteChanged();
             }
         }
         public ICommand AddServiceCommand { get; }
         public ICommand RemoveServiceCommand { get; }
         public ICommand EditServiceCommand { get; }
-        public event Action<ServiceViewModel>? EditRequested;
+        public event Action<ServiceListModel>? EditRequested;
         public int ServicesCreated => Services.Count;
         public int CurrentActiveServices => Services.Count(s => s.IsActive);
 
@@ -264,7 +66,7 @@ namespace DesktopApplicationTemplate.UI.ViewModels
             NetworkConfig = networkConfig;
             _ = NetworkConfig.LoadAsync();
             _networkService.ConfigurationChanged += (_, cfg) => ApplyNetworkConfiguration(cfg);
-            ServiceViewModel.ResolveService = (type, name) =>
+            ServiceListModel.ResolveService = (type, name) =>
                 Services.FirstOrDefault(s =>
                     s.ServiceType.Equals(type, StringComparison.OrdinalIgnoreCase) &&
                     s.DisplayName.Split(" - ").Last().Equals(name, StringComparison.OrdinalIgnoreCase));
@@ -275,7 +77,7 @@ namespace DesktopApplicationTemplate.UI.ViewModels
             }
             AddServiceCommand = new RelayCommand(AddService);
             RemoveServiceCommand = new AsyncRelayCommand(RemoveSelectedServiceAsync, () => SelectedService != null);
-            EditServiceCommand = new RelayCommand<ServiceViewModel?>(EditService, svc => svc != null);
+            EditServiceCommand = new RelayCommand<ServiceListModel?>(EditService, svc => svc != null);
             FilteredServices = CollectionViewSource.GetDefaultView(Services);
             Filters.PropertyChanged += (_, __) => ApplyFilters();
             LoadServices();
@@ -306,7 +108,7 @@ namespace DesktopApplicationTemplate.UI.ViewModels
         }
 
         public event Action? AddServiceRequested;
-        private void EditService(ServiceViewModel? service)
+        private void EditService(ServiceListModel? service)
         {
             var target = service ?? SelectedService;
             if (target != null)
@@ -397,7 +199,7 @@ namespace DesktopApplicationTemplate.UI.ViewModels
             var existing = ServicePersistence.Load(_logger);
             foreach (var info in existing.OrderBy(i => i.Order))
             {
-                var svc = new ServiceViewModel
+                var svc = new ServiceListModel
                 {
                     DisplayName = info.DisplayName,
                     ServiceType = info.ServiceType,
@@ -428,7 +230,7 @@ namespace DesktopApplicationTemplate.UI.ViewModels
         {
             FilteredServices.Filter = obj =>
             {
-                if (obj is not ServiceViewModel svc)
+                if (obj is not ServiceListModel svc)
                     return false;
 
                 if (!string.IsNullOrWhiteSpace(Filters.NameFilter) &&
@@ -448,7 +250,7 @@ namespace DesktopApplicationTemplate.UI.ViewModels
             FilteredServices.Refresh();
         }
 
-        public void OnServiceLogAdded(ServiceViewModel svc, LogEntry entry)
+        public void OnServiceLogAdded(ServiceListModel svc, LogEntry entry)
         {
             AllLogs.Insert(0, entry);
             if (svc.ServiceType != "CSV Creator" && Services.Any(s => s.ServiceType == "CSV Creator"))
