+++ DesktopApplicationTemplate.UI/Views/HttpServiceView.xaml.cs
@@ -43,7 +43,7 @@ namespace DesktopApplicationTemplate.UI.Views
             _viewModel.Logger = _logger;
         }
 
-        public void SetServiceContext(ServiceViewModel service)
+        public void SetServiceContext(ServiceListModel service)
         {
             LogView.DataContext = new ServiceLogViewModel(service.DisplayName, service.ServiceType, service.Logs);
         }
