+++ docs/CHANGELOG.md
@@ -153,7 +153,7 @@
 - Advanced TCP options support script-based message transformation with input and output previews.
 
 #### Changed
-- Registered transient TCP view models and bound `TcpServiceOptions` configuration; each `ServiceViewModel` retains its own options.
+- Registered transient TCP view models and bound `TcpServiceOptions` configuration; each `ServiceListModel` retains its own options.
 - Service persistence saves and restores `TcpServiceOptions` for TCP services.
 - Replaced `TcpServiceView` with `TcpServiceMessagesView` hosting a reusable message table.
